requirements.txt
backcall==0.2.0
decorator==5.0.9
Django==2.2
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==4.6.1
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.13.1
Markdown==3.3.4
mysqlclient==2.0.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.7.0
Pygments==2.9.0
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.10.0.0
wcwidth==0.2.5
zipp==3.5.0
